Distance Matrix with updated intermediate nodes K0-3
We will check paths from i to j using one vertex at a time (0 to 3) as an intermediate vertex. Following Floyd Warshall’s algorithm, (D[i][k] + D[k][j]) if (A[i][j] > A[i][k] + A[k][j]  every time  the distance between a pair of nodes through an intermediate vertix is lesser than the direct distance between the same two nodes, we will replace the value in the matrix for the smaller value.